+++ src/test/css/deprecated-css-custom-property-syntax_test.ts
@@ -14,40 +14,39 @@
 
 import {assert} from 'chai';
 import * as path from 'path';
-import {Analyzer, applyEdits, FSUrlLoader, makeParseLoader} from 'polymer-analyzer';
+import {Analyzer} from 'polymer-analyzer';
 
 import {Linter} from '../../linter';
 import {registry} from '../../registry';
-import {WarningPrettyPrinter} from '../util';
+import {assertExpectedFixes, WarningPrettyPrinter} from '../util';
 
 const fixtures_dir = path.join(__dirname, '..', '..', '..', 'test');
+const ruleId = `deprecated-css-custom-property-syntax`;
 
-suite('deprecated-css-custom-property-syntax', () => {
+suite(ruleId, () => {
   let analyzer: Analyzer;
   let warningPrinter: WarningPrettyPrinter;
   let linter: Linter;
 
   setup(() => {
-    analyzer = new Analyzer({urlLoader: new FSUrlLoader(fixtures_dir)});
+    analyzer = Analyzer.createForDirectory(fixtures_dir);
     warningPrinter = new WarningPrettyPrinter();
-    linter = new Linter(
-        registry.getRules(['deprecated-css-custom-property-syntax']), analyzer);
+    linter = new Linter(registry.getRules([ruleId]), analyzer);
   });
 
   test('works in the trivial case', async() => {
-    const warnings = await linter.lint([]);
+    const {warnings} = await linter.lint([]);
     assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
-    const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
+    const {warnings} =
+        await linter.lint(['perfectly-fine/polymer-element.html']);
     assert.deepEqual([...warnings], []);
   });
 
   test('warns for the proper cases and with the right messages', async() => {
-    const warnings = await linter.lint([
-      'deprecated-css-custom-property-syntax/deprecated-css-custom-property-syntax.html'
-    ]);
+    const {warnings} = await linter.lint([`${ruleId}/${ruleId}.html`]);
     assert.deepEqual(warningPrinter.prettyPrint(warnings), [
       `
     color: var(--red, --blue);
@@ -72,16 +71,10 @@ suite('deprecated-css-custom-property-syntax', () => {
   });
 
   test('applies automatic-safe fixes', async() => {
-    const warnings = await linter.lint(
-        ['deprecated-css-custom-property-syntax/before-fixes.html']);
-    const edits = warnings.filter((w) => w.fix).map((w) => w.fix!);
-    const loader = makeParseLoader(analyzer, warnings.analysis);
-    const result = await applyEdits(edits, loader);
-    assert.deepEqual(
-        result.editedFiles.get(
-            'deprecated-css-custom-property-syntax/before-fixes.html'),
-        (await loader('deprecated-css-custom-property-syntax/after-fixes.html'))
-            .contents);
-    assert.deepEqual(result.incompatibleEdits, []);
+    await assertExpectedFixes(
+        linter,
+        analyzer,
+        `${ruleId}/before-fixes.html`,
+        `${ruleId}/after-fixes.html`);
   });
 });