+++ src/test/html/undefined-elements_test.ts
@@ -14,7 +14,7 @@
 
 import * as assert from 'assert';
 import * as path from 'path';
-import {Analyzer, FSUrlLoader} from 'polymer-analyzer';
+import {Analyzer} from 'polymer-analyzer';
 
 import {Linter} from '../../linter';
 import {registry} from '../../registry';
@@ -28,23 +28,24 @@ suite('undefined-elements', () => {
   let linter: Linter;
 
   setup(() => {
-    analyzer = new Analyzer({urlLoader: new FSUrlLoader(fixtures_dir)});
+    analyzer = Analyzer.createForDirectory(fixtures_dir);
     warningPrinter = new WarningPrettyPrinter();
     linter = new Linter(registry.getRules(['undefined-elements']), analyzer);
   });
 
   test('works in the trivial case', async() => {
-    const warnings = await linter.lint([]);
+    const {warnings} = await linter.lint([]);
     assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
-    const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
+    const {warnings} =
+        await linter.lint(['perfectly-fine/polymer-element.html']);
     assert.deepEqual([...warnings], []);
   });
 
   test('finds undefined element references', async() => {
-    const warnings =
+    const {warnings} =
         await linter.lint(['undefined-elements/undefined-elements.html']);
 
     assert.deepEqual(warningPrinter.prettyPrint(warnings), [`