+++ src/test/polymer/element-before-dom-module_test.ts
@@ -16,7 +16,6 @@ import '../../rules';
 import {assert} from 'chai';
 import * as path from 'path';
 import {Analyzer} from 'polymer-analyzer';
-import {FSUrlLoader} from 'polymer-analyzer/lib/url-loader/fs-url-loader';
 
 import {Linter} from '../../linter';
 import {registry} from '../../registry';
@@ -30,24 +29,25 @@ suite('element-before-dom-module', () => {
   let linter: Linter;
 
   setup(() => {
-    analyzer = new Analyzer({urlLoader: new FSUrlLoader(fixtures_dir)});
+    analyzer = Analyzer.createForDirectory(fixtures_dir);
     warningPrinter = new WarningPrettyPrinter();
     linter =
         new Linter(registry.getRules(['element-before-dom-module']), analyzer);
   });
 
   test('works in the trivial case', async() => {
-    const warnings = await linter.lint([]);
+    const {warnings} = await linter.lint([]);
     assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
-    const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
+    const {warnings} =
+        await linter.lint(['perfectly-fine/polymer-element.html']);
     assert.deepEqual([...warnings], []);
   });
 
   test('warns for the proper cases', async() => {
-    const warnings = await linter.lint(
+    const {warnings} = await linter.lint(
         ['element-before-dom-module/element-before-dom-module.html']);
     assert.deepEqual(warningPrinter.prettyPrint(warnings), [
       `