+++ src/test/polymer/set-unknown-attribute_test.ts
@@ -16,7 +16,7 @@ import '../../rules';
 
 import {assert} from 'chai';
 import * as path from 'path';
-import {Analyzer, FSUrlLoader} from 'polymer-analyzer';
+import {Analyzer} from 'polymer-analyzer';
 
 import {Linter} from '../../linter';
 import {registry} from '../../registry';
@@ -30,23 +30,24 @@ suite('set-unknown-attribute', () => {
   let linter: Linter;
 
   setup(() => {
-    analyzer = new Analyzer({urlLoader: new FSUrlLoader(fixtures_dir)});
+    analyzer = Analyzer.createForDirectory(fixtures_dir);
     warningPrinter = new WarningPrettyPrinter();
     linter = new Linter(registry.getRules(['set-unknown-attribute']), analyzer);
   });
 
   test('works in the trivial case', async() => {
-    const warnings = await linter.lint([]);
+    const {warnings} = await linter.lint([]);
     assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
-    const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
+    const {warnings} =
+        await linter.lint(['perfectly-fine/polymer-element.html']);
     assert.deepEqual([...warnings], []);
   });
 
   test('warns at the right times', async() => {
-    const warnings =
+    const {warnings} =
         await linter.lint(['set-unknown-attribute/when-to-warn.html']);
     assert.deepEqual(warningPrinter.prettyPrint(warnings), [
       `
@@ -68,7 +69,7 @@ suite('set-unknown-attribute', () => {
   });
 
   test('gives helpful warning messages', async() => {
-    const warnings =
+    const {warnings} =
         await linter.lint(['set-unknown-attribute/warning-messages.html']);
     assert.deepEqual(warnings.map((w) => w.message), [
       'test-elem elements do not have a property named class. Consider instead:  class$',