+++ src/test/polymer/unbalanced-delimiters_test.ts
@@ -16,7 +16,7 @@ import '../../rules';
 
 import {assert} from 'chai';
 import * as path from 'path';
-import {Analyzer, FSUrlLoader} from 'polymer-analyzer';
+import {Analyzer} from 'polymer-analyzer';
 
 import {Linter} from '../../linter';
 import {registry} from '../../registry';
@@ -30,24 +30,25 @@ suite('unbalanced-polymer-delimiters', () => {
   let linter: Linter;
 
   setup(() => {
-    analyzer = new Analyzer({urlLoader: new FSUrlLoader(fixtures_dir)});
+    analyzer = Analyzer.createForDirectory(fixtures_dir);
     warningPrinter = new WarningPrettyPrinter();
     linter = new Linter(
         registry.getRules(['unbalanced-polymer-delimiters']), analyzer);
   });
 
   test('works in the trivial case', async() => {
-    const warnings = await linter.lint([]);
+    const {warnings} = await linter.lint([]);
     assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
-    const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
+    const {warnings} =
+        await linter.lint(['perfectly-fine/polymer-element.html']);
     assert.deepEqual([...warnings], []);
   });
 
   test('warns for the proper cases', async() => {
-    const warnings =
+    const {warnings} =
         await linter.lint(['unbalanced-delimiters/unbalanced-delimiters.html']);
     assert.deepEqual(warningPrinter.prettyPrint(warnings), [
       `